feat: 기본적인 API 추가

+++ module-api/src/main/java/com/example/api/domain/member/model/MemberCreateRequest.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 @Builder
 public record MemberCreateRequest(
         String name,
-        int age
+        byte age
 ) {
     public Member toEntity() {
         return Member.builder()
